+++ docs/changelog.rst
@@ -13,6 +13,12 @@ Added:
   information displayed by ``:metadata``. Thanks
   `@jeanggi90 <https://github.com/jeanggi90>`_!
 
+Changed:
+^^^^^^^^
+
+* The ``=`` key can now be bound using ``<equal>``. Using the raw ``=`` character is not
+  possible in ``keys.conf`` as it is treated as separator much like ``:``.
+
 
 v0.7.0 (2020-05-17)
 -------------------